Abstract most parsing logic out into a util function
