+++ src/library/helpers/getWords.ts
@@ -1,3 +1,4 @@
+/* eslint-disable dot-notation */
 import { API_URL } from "@env";
 
 import type { Dispatch, SetStateAction } from "react";
@@ -6,6 +7,8 @@ import type {
   FileInfo,
   LoadingData,
 } from "../../types/dataTypes";
+
+import parseSSEtoObject from "../utils/parseSSEtoObject";
 
 // If platform is web, a File object is supplied, otherwise, an object containing file details is supplied.
 
@@ -33,49 +36,19 @@ const getWords = async (
       const newData =
         test === undefined ? response : response.replace(test, "");
 
-      const responseArray = newData
-        .trim()
-        .split(/\n\n/)
-        .filter((str: string) => str);
+      const formattedData = parseSSEtoObject(newData);
 
-      const responseObjectArray = responseArray.map((data: string) => {
-        return data.split(/\n/).map((a) => {
-          const splitArray = a.split(/(?<=^[^:]*):/);
-          const key = splitArray[0]?.trim();
-          const value = splitArray[1]?.trim();
-
-          if (key === undefined) return;
-
-          const object = { [key]: value };
-          return object;
-        });
-      });
-
-      const usefulData = responseObjectArray.filter(
-        (
-          b: {
-            [key: string]: string;
-          }[]
-        ) => {
-          // eslint-disable-next-line dot-notation
-          const retryCheck = b.find((obj) => !obj["retry"]); // Eslint disabled as currently there is an ESlint typescript config regarding array accessing.
-          const endCheck = !b.find((obj) => Object.values(obj).includes("end"));
-          if (retryCheck && endCheck) return true;
-          else return false;
-        }
-      );
-
-      const flattenedData = usefulData.flat();
-      const obj1 = flattenedData[0];
-      const obj2 = flattenedData[1];
-      const formattedData = { ...obj1, ...obj2 };
       test = response;
 
-      if (formattedData.event === "loading") {
-        loadingSetter({ loading: true, message: formattedData.data });
+      if (formattedData["event"] === "loading") {
+        loadingSetter({
+          loading: true,
+          message: formattedData["data"] ?? "Loading",
+        });
       }
-      if (formattedData.event === "result") {
-        dataSetter(JSON.parse(formattedData.data));
+      if (formattedData["event"] === "result") {
+        if (formattedData["data"])
+          dataSetter(JSON.parse(formattedData["data"]));
         stream.abort();
       }
     });
